+++ Models/TransactionModel.cs
@@ -8,7 +8,7 @@ public class TransactionModel
     public string? type { get; set; } // return, spend, income
     public string? description { get; set; }
     public string? category { get; set; } // dividend, interest, capital gain, deposit, withdrawal
-    public int? value { get; set; }
+    public decimal? value { get; set; }
     public string? currencyCode { get; set; }
     public DateTime? createdOn { get; set; }
     public DateTime? updatedOn { get; set; }
